+++ src/pages/NewRoom.tsx
@@ -6,8 +6,11 @@ import illustrationImg from '../images/illustration.svg'
 import logo from "../images/logo.svg"
 
 import {Button} from "../components/Button"
+import { useAuth } from '../hooks/useAuth'
 
 export function NewRoom() {
+  const {user} = useAuth()
+  console.log(user?.name)
   return(
 
     <div id="page-auth">
@@ -21,6 +24,7 @@ export function NewRoom() {
       <main>
         <div className="main-content">
           <img src={logo} alt="LetMeAsk" />
+          <h1>{user?.name}</h1>
           <h2>Criar uma nova sala</h2>
           <form action="">
             <input
